Split Android instrumentation test packages
