+++ apps/android/app/src/androidTest/java/com/flashcardsopensourceapp/app/livesmoke/support/LiveSmokeScenarioCardHelpers.kt
@@ -7,9 +7,9 @@ import androidx.compose.ui.test.onAllNodesWithText
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performTextInput
 import androidx.compose.ui.test.performTextReplacement
-import com.flashcardsopensourceapp.app.RepositorySeedCard
-import com.flashcardsopensourceapp.app.RepositorySeedScenario
-import com.flashcardsopensourceapp.app.createRepositorySeedExecutor
+import com.flashcardsopensourceapp.app.support.RepositorySeedCard
+import com.flashcardsopensourceapp.app.support.RepositorySeedScenario
+import com.flashcardsopensourceapp.app.support.createRepositorySeedExecutor
 import com.flashcardsopensourceapp.app.livesmoke.diagnostics.clickContentDescription
 import com.flashcardsopensourceapp.app.livesmoke.diagnostics.clickNode
 import com.flashcardsopensourceapp.app.livesmoke.diagnostics.clickTag
