+++ apps/android/app/src/androidTest/java/com/flashcardsopensourceapp/app/marketing/screenshots/MarketingAllScreenshotsScript.kt
@@ -2,7 +2,7 @@ package com.flashcardsopensourceapp.app.marketing.screenshots
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.flashcardsopensourceapp.app.ManualOnlyAndroidTest
-import com.flashcardsopensourceapp.app.createRepositorySeedExecutor
+import com.flashcardsopensourceapp.app.support.createRepositorySeedExecutor
 import kotlinx.coroutines.runBlocking
 import org.junit.Assert.assertTrue
 import org.junit.Rule
